x/gold/types/message_transfer_to_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"m25/testutil/sample"
)

func TestMsgTransferTo_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgTransferTo
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgTransferTo{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgTransferTo{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
